# artifacts/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "PT2 - Employees \u0026 Tasks API",
                 "description":  "API to manage employees and tasks. Includes CRUD and basic filters.",
                 "version":  "0.1.0"
             },
    "paths":  {
                  "/employees":  {
                                     "post":  {
                                                  "summary":  "Create Employee",
                                                  "operationId":  "create_employee_employees_post",
                                                  "parameters":  [
                                                                     {
                                                                         "name":  "authorization",
                                                                         "in":  "header",
                                                                         "required":  false,
                                                                         "schema":  {
                                                                                        "anyOf":  [
                                                                                                      {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                                      {
                                                                                                          "type":  "null"
                                                                                                      }
                                                                                                  ],
                                                                                        "title":  "Authorization"
                                                                                    }
                                                                     }
                                                                 ],
                                                  "requestBody":  {
                                                                      "required":  true,
                                                                      "content":  {
                                                                                      "application/json":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/EmployeeCreate"
                                                                                                                          }
                                                                                                           }
                                                                                  }
                                                                  },
                                                  "responses":  {
                                                                    "201":  {
                                                                                "description":  "Successful Response",
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/EmployeeRead"
                                                                                                                                    }
                                                                                                                     }
                                                                                            }
                                                                            },
                                                                    "422":  {
                                                                                "description":  "Validation Error",
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                    }
                                                                                                                     }
                                                                                            }
                                                                            }
                                                                }
                                              },
                                     "get":  {
                                                 "summary":  "List Employees",
                                                 "operationId":  "list_employees_employees_get",
                                                 "parameters":  [
                                                                    {
                                                                        "name":  "skip",
                                                                        "in":  "query",
                                                                        "required":  false,
                                                                        "schema":  {
                                                                                       "type":  "integer",
                                                                                       "default":  0,
                                                                                       "title":  "Skip"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "limit",
                                                                        "in":  "query",
                                                                        "required":  false,
                                                                        "schema":  {
                                                                                       "type":  "integer",
                                                                                       "maximum":  1000,
                                                                                       "default":  100,
                                                                                       "title":  "Limit"
                                                                                   }
                                                                    }
                                                                ],
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "Successful Response",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "type":  "array",
                                                                                                                                       "items":  {
                                                                                                                                                     "$ref":  "#/components/schemas/EmployeeRead"
                                                                                                                                                 },
                                                                                                                                       "title":  "Response List Employees Employees Get"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                                   "422":  {
                                                                               "description":  "Validation Error",
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           }
                                                               }
                                             }
                                 },
...
                  "/tasks":  {
                                 "post":  {
                                              "summary":  "Create Task",
                                              "operationId":  "create_task_tasks_post",
                                              "parameters":  [
                                                                 {
                                                                     "name":  "authorization",
                                                                     "in":  "header",
                                                                     "required":  false,
                                                                     "schema":  {
                                                                                    "anyOf":  [
                                                                                                  {
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                                  {
                                                                                                      "type":  "null"
                                                                                                  }
                                                                                              ],
                                                                                    "title":  "Authorization"
                                                                                }
                                                                 }
                                                             ],
                                              "requestBody":  {
                                                                  "required":  true,
                                                                  "content":  {
                                                                                  "application/json":  {
                                                                                                           "schema":  {
                                                                                                                          "$ref":  "#/components/schemas/TaskCreate"
                                                                                                                      }
                                                                                                       }
                                                                              }
                                                              },
                                              "responses":  {
                                                                "201":  {
                                                                            "description":  "Successful Response",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/TaskRead"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                                "422":  {
                                                                            "description":  "Validation Error",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        }
                                                            }
                                          },
                                 "get":  {
                                             "summary":  "List Tasks",
                                             "operationId":  "list_tasks_tasks_get",
                                             "parameters":  [
                                                                {
                                                                    "name":  "skip",
                                                                    "in":  "query",
                                                                    "required":  false,
                                                                    "schema":  {
                                                                                   "type":  "integer",
                                                                                   "default":  0,
                                                                                   "title":  "Skip"
                                                                               }
                                                                },
                                                                {
                                                                    "name":  "limit",
                                                                    "in":  "query",
                                                                    "required":  false,
                                                                    "schema":  {
                                                                                   "type":  "integer",
                                                                                   "maximum":  1000,
                                                                                   "default":  100,
                                                                                   "title":  "Limit"
                                                                               }
                                                                },
                                                                {
                                                                    "name":  "employee_id",
                                                                    "in":  "query",
                                                                    "required":  false,
                                                                    "schema":  {
                                                                                   "anyOf":  [
                                                                                                 {
                                                                                                     "type":  "integer"
                                                                                                 },
                                                                                                 {
                                                                                                     "type":  "null"
                                                                                                 }
                                                                                             ],
                                                                                   "title":  "Employee Id"
                                                                               }
                                                                }
                                                            ],
                                             "responses":  {
                                                               "200":  {
                                                                           "description":  "Successful Response",
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "type":  "array",
                                                                                                                                   "items":  {
                                                                                                                                                 "$ref":  "#/components/schemas/TaskRead"
                                                                                                                                             },
                                                                                                                                   "title":  "Response List Tasks Tasks Get"
                                                                                                                               }
                                                                                                                }
                                                                                       }
                                                                       },
                                                               "422":  {
                                                                           "description":  "Validation Error",
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                               }
                                                                                                                }
                                                                                       }
                                                                       }
                                                           }
                                         }
                             },
...
                  "/token":  {
                                 "post":  {
                                              "summary":  "Login For Access Token",
                                              "description":  "Simple token endpoint: POST {\"username\":\"admin\", \"password\":\"secret\"} returns a JWT.\nThe expected credentials can be set via env `API_USER`/`API_PASS` (defaults: admin/secret).",
                                              "operationId":  "login_for_access_token_token_post",
                                              "requestBody":  {
                                                                  "content":  {
                                                                                  "application/json":  {
                                                                                                           "schema":  {
                                                                                                                          "type":  "object",
                                                                                                                          "title":  "Credentials"
                                                                                                                      }
                                                                                                       }
                                                                              },
                                                                  "required":  true
                                                              },
                                              "responses":  {
                                                                "200":  {
                                                                            "description":  "Successful Response",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {

                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                                "422":  {
                                                                            "description":  "Validation Error",
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        }
                                                            }
                                          }
                             },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                   }
                   }
}